修改class名
