+++ app/src/main/assets/xposed_init
@@ -1,1 +1,1 @@
-com.example.installhook.hook
+com.example.installhook.MainActivity
